+++ CHANGELOG.md
@@ -5,6 +5,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+- Functionality for renaming a file (#193)
+- Styling for a secondary button (#193)
+
+### Changed
+
+- Updated modal styling (#193)
+
 ## [0.4.0]
 
 ### Added